Fix bug with messages not updating when Chat tab is not opened
